requirements.txt

streamlit
langchain-core
langchain-community